Only tick a controller if parent entity isn't classified as dead, their position isn't null or their world isn't null. This may seem redundant but we still have cases where a controller is being ticked.
Evidence of this occurs in the current Vanilla build as of this commit and
the prior engine build. In-game simply break a block, let the item entity
lay on the ground, and log out. When the region unloads, the Item's tick
code will still fire and inccur a null when it tries to get a block from
its region.

This is probably overkill (this check) but its the ultimate sanity check
against ticking controllers when they shouldn't be. Feel free to erase
this if an alternative fix is implemented.

Signed-off-by: Chris Sanders <NinjaZidane@hotmail.com>