Cleanup the collision process. Now PreResolve only returns true if the position, world, or controller (live values) are not null. SpoutRegion only invokes resolve if pre-resolve isn't returned true for the entity and only invokes the method after all entities have been ticked for the current tick.
Within resolve, the biggest change occurs with only resolving collisions
if they have a non null collision model. Finally, EntityMoveEvent is fired
if and only if the snapshot transform doesn't equal the offset adjustment
of the collision or the live transform value. This makes sure that an
entity that moves, collides, and is offsetted back to the snapshot
position doesn't fire an EntityMoveEvent.

Seeing as preresolve doesn't return true unless a valid position and world
is true, we are safe to copy over the live transform to its snapshot for
the end of it.

Signed-off-by: Chris Sanders <NinjaZidane@hotmail.com>