Set restitution to 0 before adding body.
Apparently updating restitution after a body is added to the simulation
has no effect. This would appear to be a bug in TeraBullet and we may be
able to fix it. For now this is true non-bouncy characteristics that lets
gravity still work. It would be worth testing the other
characteristics...we may have to hotswap bodies in/out of the simulation
to avoid this issue for good.

@Afforess @Raphfrk
Signed-off-by: Chris Sanders <zidane@outlook.com>