Implemented most of scene/updated engine.
I am going to attempt the hotswapping technique to dealing with updating
characteristics of a body in a simulation: as it sounds it is
removing/setting object back in. I'll see if I can clear pair cache as
well (so I can remove the hotswapping).

I've also added a few TODOs for you @Afforess .
Signed-off-by: Chris Sanders <zidane@outlook.com>